Restructured the project and added CMakeLists
