+++ include/c_nn/Functions.hpp
