+++ include/c_nn/Functions/Cos.hpp
