+++ include/c_nn/Functions/Log.hpp
