+++ include/c_nn/Functions/Sin.hpp
