+++ libs/CMakeLists.txt
@@ -1,0 +1,3 @@
+# set(CMAKE_CXX_STANDARD 17)
+project(c_nn LANGUAGES CXX)
+# add_subdirectory(googletest)
