+++ src/CMakeLists.txt
@@ -1,0 +1,10 @@
+project(c_nn LANGUAGES CXX)
+
+file(GLOB functions_SRC
+    "Functions/*.cpp"
+)
+add_library(c_nn ${functions_SRC})
+
+target_include_directories(c_nn
+    PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include)
