+++ tests/CMakeLists.txt
@@ -1,0 +1,35 @@
+project(c_nn LANGUAGES CXX)
+
+include(FetchContent)
+
+FetchContent_Declare(
+    googletest
+    URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
+)
+
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+
+FetchContent_GetProperties(googletest)
+
+if(NOT googletest_POPULATED)
+    FetchContent_Populate(googletest)
+    add_subdirectory(${googletest_SOURCE_DIR} ${googletest_BINARY_DIR} EXCLUDE_FROM_ALL)
+endif()
+
+
+# Define tests
+enable_testing()
+
+add_executable(tests
+    pol_test.cpp
+)
+
+target_link_libraries(
+    tests
+    gtest_main
+    c_nn
+)
+
+include(GoogleTest)
+gtest_discover_tests(tests)
